ddtrace/tracer: add the WithEnv StartOption to the tracer. (#529)
The way to set the env with the go API is to set a global tag with the key "env" when starting the tracer. This is confusing and badly documented.

This commit adds a StartOption called WithEnv that will add the global tag, making it more intuitive to set the APM environment name.