ddtrace/tracer: optimize baggage item handling. (#611)
When creating a span, ForEachBaggageItem is called, which acquires a
mutex to check for and copy baggage items. Often no baggage is present,
and the expensive mutex Lock() and Unlock() are a waste. This commit
adds an atomic int32 which is set whenever the first baggage item is set,
and checked in ForEachBaggageItem to see if any baggage is present
before doing the more expensive mutex Lock(). It also checks the int
before locking in the baggageItem() function.