contrib/database/sql: add query_type span tag (#707)
This change is to add the query type to the span tags (previously resource). This will help differentiate the spans whenever a `query` has been provided to `tryTrace` (e.g. Prepare and Query for a prepared statement). If a `query` has been provided then the `resource.name` gets set to the query and not the resource.

Extends #270